+++ src/modules/question/controllers/question/get-by-question-controller.js
@@ -1,4 +1,4 @@
-import { Question } from '../models/question'
+import { Question } from '../../models/question'
 
 export default class GetByQuestionController {
   static async handle(req, res) {
