--- src/modules/questions/routes/answer-routes.js
@@ -1,9 +1,0 @@
-import { Router } from 'express'
-
-const router = Router()
-
-router.get('/', (req, res) => {
-  res.json({ respostas: 'ok' })
-})
-
-export default router
