--- src/modules/questions/routes/question-routes.js
@@ -1,22 +1,0 @@
-import { Router } from 'express'
-
-import GetAllQuestionController from '../controllers/get-all-question-controller'
-import AddQuestionController from '../controllers/add-question-controller'
-import GetByQuestionController from '../controllers/get-by-question-controller'
-import DeleteQuestionController from '../controllers/delete-question-controller'
-import UpdateQuestionController from '../controllers/update-question-controller'
-
-
-const router = Router()
-
-router.get('/', GetAllQuestionController.handle)
-
-router.get('/:id', GetByQuestionController.handle)
-
-router.post('/', AddQuestionController.handle)
-
-router.delete('/:id', DeleteQuestionController.handle)
-
-router.patch('/:id', UpdateQuestionController.handle)
-
-export default router
